add Hero component, add refetch features for Monitor component
